update runtime to run loop and capture trace
