+++ gitm/api.py
@@ -6,6 +6,7 @@
 
 from __future__ import annotations
 
+from collections.abc import Callable
 from typing import Any
 
 from gitm.scheduler import LoopConfig, run_loop
@@ -18,6 +19,7 @@ def optimize(
     budget: str = "24h",
     target: float = 0.15,
     scratch: str | None = None,
+    workload_runner: Callable[[], dict[str, Any]] | None = None,
 ) -> dict[str, Any]:
     """Run the autonomous 24-hour optimization loop and return a report.
 
@@ -26,6 +28,11 @@ def optimize(
     path. ``budget`` and ``target`` follow the SKU contract: a verified floor
     of ``target`` fraction improvement within ``budget`` wall time, or a
     qualification-gate diagnostic explaining why the floor was not committed.
+
+    ``workload_runner`` optionally supplies an explicit zero-arg callable that
+    launches the workload's GPU work; it runs inside the capture window. When
+    omitted, the loop resolves ``workload`` against the registry in
+    :mod:`gitm.workloads`.
     """
     cfg = LoopConfig(
         engine=engine,
@@ -33,5 +40,6 @@ def optimize(
         budget=budget,
         target=target,
         scratch=scratch,
+        workload_runner=workload_runner,
     )
     return run_loop(cfg)
